Fix websockets dependency inadvertently becoming non-optional
Since WebsocketLLMService was introduced, core llm_service.py imported
the optional `websockets` package unconditionally. So a missing
`websockets` broke importing every LLM service with a raw ImportError —
short-circuiting the try/except guard each service wraps its optional
imports in to point users at the right `pipecat-ai[...]`
extra (e.g. `[openai]`).

Move WebsocketLLMService and WebsocketReconnectedError (used only by
OpenAIResponsesLLMService) to a new websocket_llm_service module that
owns the dependency. llm_service.py re-exports them lazily via
__getattr__ to keep the released import path working.
